requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
openai
chromadb
python-dotenv
youtube-transcript-api
pytube
